document-processing/ui/requirements.txt
streamlit==1.34.0
pillow==10.4.0
types-requests
python-dotenv==1.0.0